+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [core] feature: glow the `search button` in all `paged GUI`.
+- [core] fix: should skip the linking of GUIs, if the `search function` returns the identical contents to current contents.